final-pjt-back/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
asttokens==2.4.1
black==24.4.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.56.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
environ==1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.35.13
packaging==24.1
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==26.0.3
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tiktoken==0.7.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.12
zipp==3.19.2
